src/backend/requirements.txt
# =============================================================================
# Codex Dominion Backend - Production Dependencies
# =============================================================================

# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# HTTP & CORS
httpx==0.26.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Data validation
email-validator==2.1.0

# Date/Time
python-dateutil==2.8.2

# Logging & Monitoring
structlog==24.1.0

# Testing (optional for production)
pytest==8.0.0
pytest-asyncio==0.23.3
